add response concern and fix propertyes controller
