+++ config/application.rb
@@ -18,5 +18,6 @@ module TabasApp
     #
     # config.time_zone = "Central Time (US & Canada)"
     # config.eager_load_paths << Rails.root.join("extras")
+    config.autoloader = :classic
   end
 end
